requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-ckeditor-5==0.2.13
django-jazzmin==3.0.0
django-parler==2.3
django-resized==1.0.2
django-rosetta==0.10.0
engineering-notation==0.10.0
idna==3.7
numpy==1.26.4
pillow==10.3.0
polib==1.2.0
python-dotenv==1.0.1
requests==2.31.0
sorl-thumbnail==12.10.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
